PASS1-576: Render camera image direct to LCD
Instead of letting LVGL render pixel-by-pixel from the camera image downsampled to 1 bit-per-pixel, we now render directly to the LCD driver buffer and update just the modified lines.

In addition, the LVGL `_lcd_rounder` callback was always setting the fully display to be updated (in addition to setting `direct_mode=1`, which ended up doing the same thing).  Change it so that only the dirty lines are push to the display.  Also optimized the bit calculation of `_lcd_set_px()` to be a table lookup instead for a small LVGL performance boost.